+++ src/plugins/dev-mode/error-messages.ts
@@ -215,6 +215,9 @@ export const ERROR_MESSAGES = {
     CRDT2: 'RxDocument.atomicUpdate() cannot be used when CRDTs are activated.',
     CRDT3: 'To use CRDTs you MUST NOT set a conflictHandler because the default CRDT conflict handler must be used',
 
+    // plugins/dexie/
+    DXE1: 'The dexie.js RxStorage does not support boolean indexes, see https://rxdb.info/rx-storage-dexie.html#boolean-index',
+
     /**
      * Should never be thrown, use this for
      * null checks etc. so you do not have to increase the