+++ src/plugins/dexie/rx-storage-dexie.ts
@@ -15,6 +15,7 @@ import {
     Query as MingoQuery
 } from 'mingo';
 import {
+    ensureNoBooleanIndex,
     getDexieSortComparator,
     RX_STORAGE_NAME_DEXIE
 } from './dexie-helper';
@@ -104,6 +105,7 @@ export class RxStorageDexie implements RxStorage<DexieStorageInternals, DexieSet
         params: RxStorageInstanceCreationParams<RxDocType, DexieSettings>
     ): Promise<RxStorageInstanceDexie<RxDocType>> {
         ensureRxStorageInstanceParamsAreCorrect(params);
+        ensureNoBooleanIndex(params.schema);
         return createDexieStorageInstance(this, params, this.settings);
     }
 }